Package name refactor for binary relations
Refactor the package name
com.github.vincentk.dedekind.relation.binary
to
com.github.vincentk.dedekind.binary.relation
