+++ dedekind-ontology/src/main/java/com/github/vincentk/dedekind/algebra/numbers/B.java
@@ -6,8 +6,8 @@ package com.github.vincentk.dedekind.algebra.numbers;
 import com.github.vincentk.dedekind.sets.Cardinality;
 import com.github.vincentk.dedekind.algebra.sets.SemiRings.Booleans;
 import com.github.vincentk.dedekind.sets.Set;
+import com.github.vincentk.dedekind.sets.binary.relation.homogeneous.Ring;
 import com.github.vincentk.dedekind.sets.ordered.TotallyOrdered;
-import com.github.vincentk.dedekind.sets.relation.binary.homogeneous.Ring;
 
 /**
  * Boolean values. Roughly speaking 0 <=> false, 1 <=> true.
