+++ dedekind-ontology/src/main/java/com/github/vincentk/dedekind/algebra/numbers/MetricSpace.java
@@ -5,7 +5,7 @@ package com.github.vincentk.dedekind.algebra.numbers;
 
 import com.github.vincentk.dedekind.algebra.sets.Fields;
 import com.github.vincentk.dedekind.sets.Set;
-import com.github.vincentk.dedekind.sets.relation.binary.homogeneous.Monoid;
+import com.github.vincentk.dedekind.sets.binary.relation.homogeneous.Monoid;
 
 /**
  * @see https://en.wikipedia.org/wiki/Metric_space
