+++ dedekind-ontology/src/main/java/com/github/vincentk/dedekind/algebra/numbers/N.java
@@ -5,8 +5,8 @@ package com.github.vincentk.dedekind.algebra.numbers;
 
 import com.github.vincentk.dedekind.sets.Cardinality;
 import com.github.vincentk.dedekind.sets.Set;
+import com.github.vincentk.dedekind.sets.binary.relation.homogeneous.SemiRing;
 import com.github.vincentk.dedekind.sets.ordered.TotallyOrdered;
-import com.github.vincentk.dedekind.sets.relation.binary.homogeneous.SemiRing;
 
 /**
  * The natural numbers.
