+++ dedekind-ontology/src/main/java/com/github/vincentk/dedekind/algebra/numbers/Number.java
@@ -4,7 +4,7 @@
 package com.github.vincentk.dedekind.algebra.numbers;
 
 import com.github.vincentk.dedekind.sets.Set;
-import com.github.vincentk.dedekind.sets.relation.binary.homogeneous.SemiRing;
+import com.github.vincentk.dedekind.sets.binary.relation.homogeneous.SemiRing;
 
 /**
  * In this context, a number is a member of a set with at least
