+++ dedekind-ontology/src/main/java/com/github/vincentk/dedekind/algebra/numbers/NumberLine.java
@@ -4,10 +4,10 @@
 package com.github.vincentk.dedekind.algebra.numbers;
 
 import com.github.vincentk.dedekind.sets.Cardinality;
+import com.github.vincentk.dedekind.sets.binary.relation.homogeneous.Ring;
+import com.github.vincentk.dedekind.sets.binary.relation.homogeneous.SemiRing;
 import com.github.vincentk.dedekind.sets.ordered.TotallyOrdered;
 import com.github.vincentk.dedekind.algebra.sets.Fields;
-import com.github.vincentk.dedekind.sets.relation.binary.homogeneous.Ring;
-import com.github.vincentk.dedekind.sets.relation.binary.homogeneous.SemiRing;
 
 /**
  * A specialization of the {@link MetricSpace} to sets of {@link Number} with a total order.
