+++ dedekind-ontology/src/main/java/com/github/vincentk/dedekind/algebra/numbers/R.java
@@ -1,7 +1,7 @@
 package com.github.vincentk.dedekind.algebra.numbers;
 
 import com.github.vincentk.dedekind.sets.Cardinality;
-import com.github.vincentk.dedekind.sets.relation.binary.homogeneous.Field;
+import com.github.vincentk.dedekind.sets.binary.relation.homogeneous.Field;
 
 /**
  * The set of real numbers.
