+++ dedekind-ontology/src/main/java/com/github/vincentk/dedekind/bilinear/OuterProduct.java
@@ -1,9 +1,9 @@
 package com.github.vincentk.dedekind.bilinear;
 
 import com.github.vincentk.dedekind.linear.matrix.Matrix;
-import com.github.vincentk.dedekind.sets.relation.binary.Bracket.Bra;
-import com.github.vincentk.dedekind.sets.relation.binary.Bracket.Ket;
-import com.github.vincentk.dedekind.sets.relation.binary.homogeneous.SemiRing;
+import com.github.vincentk.dedekind.sets.binary.relation.Bracket.Bra;
+import com.github.vincentk.dedekind.sets.binary.relation.Bracket.Ket;
+import com.github.vincentk.dedekind.sets.binary.relation.homogeneous.SemiRing;
 
 /**
  * Outer product / tensor product implementation.
