+++ dedekind-ontology/src/main/java/com/github/vincentk/dedekind/families/Enumeration.java
@@ -2,7 +2,7 @@ package com.github.vincentk.dedekind.families;
 
 import java.util.Optional;
 
-import com.github.vincentk.dedekind.sets.relation.binary.Pair;
+import com.github.vincentk.dedekind.sets.binary.relation.Pair;
 
 @FunctionalInterface
 public interface Enumeration<T, E extends Enumeration<T, E>>
