+++ dedekind-ontology/src/main/java/com/github/vincentk/dedekind/linear/Basis.java
@@ -4,8 +4,8 @@
 package com.github.vincentk.dedekind.linear;
 
 import com.github.vincentk.dedekind.sets.Cardinality;
-import com.github.vincentk.dedekind.sets.relation.binary.Vector;
-import com.github.vincentk.dedekind.sets.relation.binary.homogeneous.Field;
+import com.github.vincentk.dedekind.sets.binary.relation.Vector;
+import com.github.vincentk.dedekind.sets.binary.relation.homogeneous.Field;
 
 /**
  * An appropriate way to model the notion of a basis would appear to
