+++ dedekind-ontology/src/test/java/com/github/vincentk/dedekind/binary/relation/homogeneous/PartialOrderTest.java
@@ -1,4 +1,4 @@
-package com.github.vincentk.dedekind.relation.binary.homogeneous;
+package com.github.vincentk.dedekind.binary.relation.homogeneous;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
