+++ dedekind-ontology/src/test/java/com/github/vincentk/dedekind/linear/matrix/MatrixTest.java
@@ -5,7 +5,7 @@ import static org.mockito.Mockito.mock;
 import org.junit.jupiter.api.Test;
 
 import com.github.vincentk.dedekind.algebra.numbers.R;
-import com.github.vincentk.dedekind.sets.relation.binary.Bracket.Ket;
+import com.github.vincentk.dedekind.sets.binary.relation.Bracket.Ket;
 
 public class MatrixTest {
